requirements.txt
fastapi
uvicorn
opencv-python
Pillow
mtcnn
numpy
setuptools
tensorflow